Break storage.
Add following tables :
* log
* account
* volumes
* balances

Now, we only write on the "log" table.
There is a set of triggers which are in charge of filling tables when a write is make on "log" table.
Thus, the hash is now on logs instead of transactions.
This way the hash take in account, metadata applied to transactions or accounts.
The automatic migration fill this log table with the existing data.

Since, the new code use some json function, the code must be compiled with "json1" build tag to enable support on sqlite.