Add redis locking strategy.
This PR allow to configure the lock strategy used by the ledger to prevent concurrent transactions.
Following flag was defined :
* --lock-strategy: memory, none, or redis (default: memory)

The 'none' strategy disable locking. The client could handle the retry itself.
The 'memory' strategy is the default and is the same as before.
The 'redis' strategy use a lock on redis.

'redis' strategy use additionnal flags :
* --lock-strategy-redis-url: Url of the redis server
* --lock-strategy-redis-duration: Duration before redis remove the lock whatever the job is terminated or not (prevent deadlock if a service crash)
* --lock-strategy-redis-retry: It is the waiting duration between two attempts to get a lock (if already take by another transaction).