feat(api): Add configurable page_size on paginated request.
This add the query parameter 'page_size' on both GET /transactions, GET /accounts, and GET /balances endpoints.
The parameter is capped to a value of 1000.
The default value still the same as before (15).

The PR also fix some tests around pagination which was working but was not doing the right thing (see api/controllers/pagination_test.go line 209).